Drop the build; ship a single self-contained src/index.js on k6
k6 runs ES modules and TypeScript directly, so the webpack + ts-loader bundle
was unnecessary. The only reason to keep it was producing a single index.js for
jslib.k6.io, which we now do by hand.

- Inline check.ts, stages.js and utils.js into a single plain-JS src/index.js
  (same 11 exports, behavior unchanged). check.ts's TypeScript types are
  stripped; the published artifact never shipped types, so this is not a
  regression.
- Remove webpack, ts-loader, babel, terser, package.json, package-lock.json,
  webpack.config.js, tsconfig.json, prettier/husky/lint-staged/yarn/npm config,
  the stale committed build/ bundle, and .gitignore.
- Repoint tests from ../build/index.js and ../src/{utils,stages}.js to
  ../src/index.js.
- Add .editorconfig and a node-free CI workflow (grafana/setup-k6-action);
  update README and AGENTS.md.
