+++ .github/workflows/test.yml
@@ -1,0 +1,24 @@
+name: Test
+
+on:
+  push:
+    branches: [master, main]
+  pull_request:
+
+jobs:
+  k6-tests:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Install k6
+        uses: grafana/setup-k6-action@v1
+
+      - name: Run k6 tests
+        run: |
+          for test in tests/*.js tests/*.ts; do
+            echo "::group::k6 run $test"
+            k6 run "$test"
+            echo "::endgroup::"
+          done
