--- .prettierignore
@@ -1,6 +1,0 @@
-# Ignore artifacts
-build
-coverage
-.prettierrc.json
-babel.config.json
-webpack.config.js
