--- .yarnrc
@@ -1,1 +1,0 @@
-ignore-scripts true
