+++ AGENTS.md
@@ -1,10 +1,11 @@
 # k6-jslib-utils
 
-JavaScript/TypeScript utility library for k6 load tests, imported by users from jslib.k6.io.
+JavaScript utility library for k6 load tests, imported by users from jslib.k6.io.
 
 ## Architecture
 
-Three source modules feed into a single webpack bundle that ships as CommonJS:
+All helpers live in a single self-contained ES module, `src/index.js`, which k6 runs
+directly — there is no build step. It is organized into three groups:
 
 - **Randomness and string helpers** produce UUIDs, random strings/ints, and extract substrings. UUID generation has two paths: a fast Math.random path and a cryptographically secure path using k6's crypto API. Callers pick via a boolean flag; the secure path is significantly slower.
 
@@ -12,13 +13,13 @@ Three source modules feed into a single webpack bundle that ships as CommonJS:
 
 - **Async-aware check wrapper** reports pass/fail through a Rate metric that shares the same well-known metric name as k6's built-in check. It handles promises by recursing on resolution, letting it work transparently in both sync and async k6 scripts.
 
-Data flow: source modules import only from k6 built-in APIs (crypto, metrics, execution). The webpack build treats all k6 imports and URLs as externals so the bundle stays self-contained.
-
-The bundled output is committed to the repo because jslib.k6.io serves it directly. Any source change requires a rebuild or the published artifact goes stale.
+`src/index.js` imports only from k6 built-in APIs (crypto, metrics, execution), so it
+is self-contained. Publishing is just copying `src/index.js` to jslib.k6.io as
+`index.js` — no bundling, no transpilation. (The `check` helper used to be TypeScript
+in `src/check.ts`; since the published artifact never shipped its types, it was
+inlined as plain JS when the build was removed.)
 
 ## Gotchas
-
-- The webpack config declares the `externals` key twice. The second declaration (a regex) silently overwrites the first (an array). Both appear intentional, but the first is dead code.
 
 - Stage index computation uses wall-clock time, which drifts from k6's internal clock if VU initialization or setup takes nontrivial time. There is no synchronization with the engine's stage progression.
