--- babel.config.json
@@ -1,11 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env"
-    ],
-    [
-      "@babel/preset-typescript"
-    ] 
-  ],
-  "plugins": [ "@babel/plugin-transform-block-scoping"]
-}
