--- package.json
@@ -1,38 +1,0 @@
-{
-  "name": "k6-jslib-utils",
-  "repository": "https://github.com/grafana/k6-jslib-utils",
-  "version": "1.3.0",
-  "packageManager": "npm@10.9.8",
-  "main": "src/index.js",
-  "devDependencies": {
-    "@babel/core": "7.29.0",
-    "@babel/plugin-transform-block-scoping": "7.28.6",
-    "@babel/preset-env": "7.29.5",
-    "@types/k6": "0.57.1",
-    "@types/webpack": "5.28.5",
-    "babel-loader": "8.4.1",
-    "clean-webpack-plugin": "4.0.0",
-    "husky": "9.1.7",
-    "lint-staged": "12.5.0",
-    "prettier": "3.8.3",
-    "terser-webpack-plugin": "5.5.0",
-    "ts-loader": "9.6.0",
-    "webpack": "5.106.2",
-    "webpack-cli": "7.0.2",
-    "webpack-glob-entries": "1.0.1"
-  },
-  "engines": {},
-  "scripts": {
-    "webpack": "webpack",
-    "test": "k6 run tests/",
-    "prepare": "husky"
-  },
-  "keywords": [
-    "k6"
-  ],
-  "author": "k6 team",
-  "license": "MIT",
-  "lint-staged": {
-    "**/*": "prettier --write --ignore-unknown"
-  }
-}
