--- test-v9.js
@@ -1,2 +1,0 @@
-// Test Husky v9 without deprecation warnings
-const badly = 'formatted'
