--- tsconfig.json
@@ -1,15 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "es5",
-        "module": "commonjs",
-        "outDir": "./dist",
-        "rootDir": "./src",
-        "strict": true,
-        "esModuleInterop": true,
-        "forceConsistentCasingInFileNames": true,
-        "declaration": true,
-        "declarationDir": "./build",
-        "allowJs": true
-    },
-    "include": ["src/**/*.ts", "src/**/*.js"]
-}
